Refrain caching map value hash codes
Avoids issues when map values are not immutable, i.e., change over
time or are lazily computed.

This changes the run-time characteristics of `hashCode()` from O(1) to
O(n), as it requires traversing the whole prefix tree.

Furthermore, it is not possible to reuse the (cached) key hash codes as
the hash code of a `java.util.Map` is defined as the sum of the entry hash
codes (i.e., `entry.getKey().hashCode() ^ entry.getValue().hashCode()`).